+++ clock_app.py
@@ -9,7 +9,7 @@ except ImportError:
     Image = None
     ImageTk = None
 
-APP_VERSION = 'V0.7'
+APP_VERSION = 'V0.8'
 
 
 class FullscreenClockApp:
@@ -686,5 +686,6 @@ def main() -> None:
 
 if __name__ == '__main__':
     main()
+